chore(realtime)!: remove old realtime client
