+++ Sources/Realtime/PresenceAction.swift
@@ -8,7 +8,7 @@
 import _Helpers
 import Foundation
 
-public struct PresenceV2: Hashable, Sendable {
+public struct Presence: Hashable, Sendable {
   /// The presence reference of the object.
   public let ref: String
 
@@ -17,7 +17,7 @@ public struct PresenceV2: Hashable, Sendable {
   public let state: JSONObject
 }
 
-extension PresenceV2: Codable {
+extension Presence: Codable {
   struct _StringCodingKey: CodingKey {
     var stringValue: String
 
@@ -68,7 +68,7 @@ extension PresenceV2: Codable {
     }
 
     meta["phx_ref"] = nil
-    self = PresenceV2(ref: presenceRef, state: meta)
+    self = Presence(ref: presenceRef, state: meta)
   }
 
   public func encode(to encoder: any Encoder) throws {
@@ -91,15 +91,15 @@ extension PresenceV2: Codable {
 
 /// Represents a presence action.
 public protocol PresenceAction: Sendable, HasRawMessage {
-  /// Represents a map of ``PresenceV2`` objects indexed by their key.
+  /// Represents a map of ``Presence`` objects indexed by their key.
   ///
   /// Your own key can be customized when creating the channel within the presence config.
-  var joins: [String: PresenceV2] { get }
+  var joins: [String: Presence] { get }
 
-  /// Represents a map of ``PresenceV2`` objects indexed by their key.
+  /// Represents a map of ``Presence`` objects indexed by their key.
   ///
   /// Your own key can be customized when creating the channel within the presence config.
-  var leaves: [String: PresenceV2] { get }
+  var leaves: [String: Presence] { get }
 }
 
 extension PresenceAction {
@@ -137,7 +137,7 @@ extension PresenceAction {
 }
 
 struct PresenceActionImpl: PresenceAction {
-  var joins: [String: PresenceV2]
-  var leaves: [String: PresenceV2]
-  var rawMessage: RealtimeMessageV2
+  var joins: [String: Presence]
+  var leaves: [String: Presence]
+  var rawMessage: RealtimeMessage
 }
